requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
Brlapi==0.6.6
caffeine==2.9.4
certifi==2018.1.18
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
ewmh==0.1.5
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
httplib2==0.9.2
idna==2.6
isort==4.3.17
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1
MarkupSafe==1.0
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
netifaces==0.10.4
numpy==1.16.2
oauth==1.0.1
olefile==0.45.1
pbr==5.1.3
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.7.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.0
pyRFC3339==1.0
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
python-xlib==0.20
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
vboxapi==1.0
virtualenv==16.4.3
wadllib==1.3.2
Werkzeug==0.15.2
wrapt==1.11.1
xkit==0.0.0
zope.interface==4.3.2
